Authentication microservice implementation
Implemented the base of the authentication microservice. So far, all endpoints are implemented (login, register employee, register customer), but only the login endpoint has been tested. The login procedure works in such a way that the user sends a request containing a username and password, and on a successful authentication response receives a JsonWebToken containing information about the issuer, timeout, his role, etc.